showing all timeslots of chosen day, behave unique
